# docs/openapi/settings-service.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "@tartware/core-service API",
    "description": "Settings catalog service APIs for tenant metadata.",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/v1/settings/categories": {
      "get": {
        "summary": "List settings categories",
        "tags": [
          "Settings Catalog"
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "default": true
            },
            "in": "query",
            "name": "active_only",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "tenant_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "is_deleted": {
                            "type": "boolean"
                          },
                          "deleted_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "deleted_by": {
                            "type": "string",
                            "maxLength": 100
                          },
                          "code": {
                            "type": "string",
                            "minLength": 2,
                            "maxLength": 64,
                            "pattern": "^[A-Z0-9_]+$"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 3,
                            "maxLength": 120
                          },
                          "description": {
                            "type": "string",
                            "maxLength": 1024
                          },
                          "icon": {
                            "type": "string",
                            "maxLength": 64
                          },
                          "color": {
                            "type": "string",
                            "maxLength": 32
                          },
                          "sort_order": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0
                          },
                          "is_active": {
                            "type": "boolean",
                            "default": true
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "metadata": {
                            "type": "object",
                            "additionalProperties": {},
                            "description": "Flexible JSONB metadata object"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "tenant_id",
                          "code",
                          "name",
                          "created_at"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "count"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/settings/sections": {
      "get": {
        "summary": "List settings sections",
        "tags": [
          "Settings Catalog"
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "default": true
            },
            "in": "query",
            "name": "active_only",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "category_id",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 64
            },
            "in": "query",
            "name": "category_code",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "tenant_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "is_deleted": {
                            "type": "boolean"
                          },
                          "deleted_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "deleted_by": {
                            "type": "string",
                            "maxLength": 100
                          },
                          "category_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "code": {
                            "type": "string",
                            "minLength": 2,
                            "maxLength": 64,
                            "pattern": "^[A-Z0-9_]+$"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 3,
                            "maxLength": 160
                          },
                          "description": {
                            "type": "string",
                            "maxLength": 2048
                          },
                          "icon": {
                            "type": "string",
                            "maxLength": 64
                          },
                          "sort_order": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0
                          },
                          "is_active": {
                            "type": "boolean",
                            "default": true
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "metadata": {
                            "type": "object",
                            "additionalProperties": {},
                            "description": "Flexible JSONB metadata object"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "tenant_id",
                          "category_id",
                          "code",
                          "name",
                          "created_at"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "type": "integer",
                          "minimum": 0
                        }
                      },
                      "required": [
                        "count"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/settings/catalog/{categoryCode}": {
      "get": {
        "summary": "Retrieve catalog metadata for a specific category",
        "tags": [
          "Settings Catalog"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "categoryCode",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/settings/values/{valueId}": {
      "patch": {
        "summary": "Update a settings value",
        "tags": [
          "Settings Catalog"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "setting_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "scope_level": {
                    "type": "string",
                    "enum": [
                      "GLOBAL",
                      "TENANT",
                      "TENANT_TEMPLATE",
                      "PROPERTY",
                      "UNIT",
                      "USER"
                    ]
                  },
                  "value": {},
                  "property_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "unit_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "user_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "PENDING",
                      "EXPIRED"
                    ]
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1024
                  },
                  "effective_from": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "effective_to": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "context": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "locked_until": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "valueId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}